+++ src/main/java/io/github/demchaav/markdown/parser/FlexmarkMarkdownParser.java
@@ -1,6 +1,7 @@
 package io.github.demchaav.markdown.parser;
 
 import com.vladsch.flexmark.ext.gfm.strikethrough.StrikethroughExtension;
+import com.vladsch.flexmark.ext.gfm.tasklist.TaskListExtension;
 import com.vladsch.flexmark.ext.tables.TablesExtension;
 import com.vladsch.flexmark.parser.Parser;
 import com.vladsch.flexmark.util.ast.Document;
@@ -23,7 +24,8 @@ public final class FlexmarkMarkdownParser {
     /** Creates a parser configured for CommonMark plus GFM strikethrough. */
     public FlexmarkMarkdownParser() {
         MutableDataSet options = new MutableDataSet();
-        options.set(Parser.EXTENSIONS, List.of(StrikethroughExtension.create(), TablesExtension.create()));
+        options.set(Parser.EXTENSIONS, List.of(
+                StrikethroughExtension.create(), TablesExtension.create(), TaskListExtension.create()));
         this.parser = Parser.builder(options).build();
     }